[color] fixed displayed trim range

+++ radio/src/gui/colorlcd/layouts/trims.cpp
@@ -25,7 +25,7 @@ void MainViewTrim::checkEvents()
 {
   Window::checkEvents();
   int8_t stickIndex = CONVERT_MODE(idx);
-  int8_t newValue = getTrimValue(mixerCurrentFlightMode, stickIndex);
+  int newValue = getTrimValue(mixerCurrentFlightMode, stickIndex);
   if (value != newValue) {
     value = newValue;
     invalidate();
